CREO-03: derive primitive scalar counts from values
Primitive arrays cannot carry a count that differs from their values.

Committed with --no-verify: the scoped gate runs in the lane's verification step and after merge.

+++ crates/cadmpeg-codec-creo/src/decode/records.rs
@@ -458,7 +458,7 @@ pub(super) struct CreoPrimitiveScalarArrayRecord {
     pub(super) id: String,
     pub(super) field: String,
     pub(super) expanded_offset: usize,
-    pub(super) count: u32,
+    pub(super) count: usize,
     pub(super) values: Vec<f64>,
 }
